+++ pallets/subspace/src/module.rs
@@ -80,7 +80,9 @@ impl ModuleChangeset {
             let floor = FloorDelegationFee::<T>::get();
             ensure!(fee >= floor, Error::<T>::InvalidMinDelegationFee);
 
-            DelegationFee::<T>::insert(netuid, &key, fee);
+            if !DelegationFee::<T>::contains_key(&key) {
+                DelegationFee::<T>::insert(&key, fee);
+            }
         }
 
         if let Some(metadata) = self.metadata {
@@ -123,7 +125,7 @@ impl<T: Config> Pallet<T> {
             name: Name::<T>::get(netuid, uid),
             address: Address::<T>::get(netuid, uid),
             metadata: Metadata::<T>::get(netuid, key),
-            delegation_fee: DelegationFee::<T>::get(netuid, key),
+            delegation_fee: DelegationFee::<T>::get(key),
             controller: key.clone(),
         }
     }
@@ -265,6 +267,9 @@ impl<T: Config> Pallet<T> {
         ValidatorPermits::<T>::insert(netuid, validator_permit);
         ValidatorTrust::<T>::insert(netuid, validator_trust);
 
+        WeightSetAt::<T>::set(netuid, uid, WeightSetAt::<T>::get(netuid, replace_uid));
+        WeightSetAt::<T>::remove(netuid, replace_uid);
+
         // SWAP WEIGHTS
         Weights::<T>::insert(netuid, uid, Weights::<T>::get(netuid, replace_uid)); // Make uid - key association.
         Weights::<T>::remove(netuid, replace_uid); // Make uid - key association.
@@ -294,15 +299,11 @@ impl<T: Config> Pallet<T> {
         Name::<T>::remove(netuid, replace_uid);
 
         // HANDLE THE DELEGATION FEE
-        DelegationFee::<T>::insert(
-            netuid,
-            &module_key,
-            DelegationFee::<T>::get(netuid, &replace_key),
-        ); // Make uid - key association.
-        DelegationFee::<T>::remove(netuid, &replace_key); // Make uid - key association.
-
-        // remove stake from old key and add to new key
-        Self::remove_stake_from_storage(&module_key);
+        if Uids::<T>::iter().all(|(_, key, _)| key != module_key) {
+            DelegationFee::<T>::remove(&module_key);
+            // Remove stake from old key and add to new key
+            Self::remove_stake_from_storage(&module_key);
+        }
 
         // 3. Remove the network if it is empty.
         let module_count = N::<T>::mutate(netuid, |v| {